novena: eeprom: Don't disable features in recovery mode
If a user sets an unsupported selection of flags, it could make it very
difficult to recover.  For example, if they disable everything, the only
way to fix it is to log in via SSH on the 100 Mbit line, or use a serial
console.  To compensate, enable all features in recovery, the way it's done
in the factory.

Signed-off-by: Sean Cross <xobs@kosagi.com>