video: ipu_disp: add timeout to int2_stat check
This fixes a hang which can happen occasionally where we load
the kernel before the monitor has sunk to the display.  We end
up waiting in this loop for an interrupt that doesn't happen.

It is safe to just break if we don't get something right away, so
we don't hang the system.